requirements.txt
Flask
gunicorn
google-cloud-storage
pillow
numpy
tensorflow
keras
